Modularization: simplified typing
